DTSCCI-2144: Simplify AllPartiesEmailGenerator
