+++ src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/claimdismissed/ClaimDismissedRespSolTwoEmailDTOGenerator.java
@@ -8,18 +8,18 @@ import uk.gov.hmcts.reform.civil.service.OrganisationService;
 @Component
 public class ClaimDismissedRespSolTwoEmailDTOGenerator extends RespSolTwoEmailDTOGenerator {
 
-    private final ClaimDismissedEmailHelper claimDismissedEmailHelper;
+    private final ClaimDismissedEmailTemplater claimDismissedEmailTemplater;
     protected static final String REFERENCE_TEMPLATE_RESPONDENT_FOR_CLAIM_DISMISSED = "claim-dismissed-respondent-notification-%s";
 
     public ClaimDismissedRespSolTwoEmailDTOGenerator(OrganisationService organisationService,
-                                                     ClaimDismissedEmailHelper claimDismissedEmailHelper) {
+                                                     ClaimDismissedEmailTemplater claimDismissedEmailHelper) {
         super(organisationService);
-        this.claimDismissedEmailHelper = claimDismissedEmailHelper;
+        this.claimDismissedEmailTemplater = claimDismissedEmailHelper;
     }
 
     @Override
     public String getEmailTemplateId(CaseData caseData) {
-        return claimDismissedEmailHelper.getTemplateId(caseData); //NOSONAR
+        return claimDismissedEmailTemplater.getTemplateId(caseData);
     }
 
     @Override
@@ -29,6 +29,6 @@ public class ClaimDismissedRespSolTwoEmailDTOGenerator extends RespSolTwoEmailDT
 
     @Override
     protected Boolean getShouldNotify(CaseData caseData) {
-        return claimDismissedEmailHelper.isValidForRespondentEmail(caseData); //NOSONAR
+        return super.getShouldNotify(caseData) && caseData.getClaimDismissedDate() != null;
     }
 }
